# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Brief": {
        "description": "El encargo de la obra.\n\n`start` es el instante en que arranca el mundo: todo lo que el brief\nestablece es cierto desde ahi, y nada anterior existe.\n\nLo que va despues de `word_tolerance` lo trae la entrevista (RF-200). Es\nopcional para que un brief escrito a mano, sin destinatario, siga valiendo.",
        "properties": {
          "dedication": {
            "default": "",
            "title": "Dedication",
            "type": "string"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/BriefEntity"
            },
            "title": "Entities",
            "type": "array"
          },
          "forbidden_themes": {
            "items": {
              "type": "string"
            },
            "title": "Forbidden Themes",
            "type": "array"
          },
          "forbidden_words": {
            "items": {
              "type": "string"
            },
            "title": "Forbidden Words",
            "type": "array"
          },
          "genre": {
            "default": "",
            "title": "Genre",
            "type": "string"
          },
          "recipient": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recipient"
              },
              {
                "type": "null"
              }
            ]
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/BriefRelation"
            },
            "title": "Relations",
            "type": "array"
          },
          "rulebook": {
            "default": "",
            "description": "DEP-02, vacio si no es deportiva",
            "title": "Rulebook",
            "type": "string"
          },
          "start": {
            "$ref": "#/components/schemas/WorldTime"
          },
          "style_guide": {
            "description": "POE-06",
            "minLength": 1,
            "title": "Style Guide",
            "type": "string"
          },
          "target_words": {
            "exclusiveMinimum": 0.0,
            "title": "Target Words",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "tone": {
            "default": "",
            "title": "Tone",
            "type": "string"
          },
          "word_tolerance": {
            "default": 0.1,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Word Tolerance",
            "type": "number"
          }
        },
        "required": [
          "title",
          "start",
          "entities",
          "style_guide",
          "target_words"
        ],
        "title": "Brief",
        "type": "object"
      },
      "BriefEntity": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "attributes": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "title": "Attributes",
            "type": "array"
          },
          "competences": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "title": "Competences",
            "type": "array"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "description": "person | place | institution | object",
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name"
        ],
        "title": "BriefEntity",
        "type": "object"
      },
      "BriefRelation": {
        "properties": {
          "kind": {
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "source": {
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "target": {
            "minLength": 1,
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target",
          "kind"
        ],
        "title": "BriefRelation",
        "type": "object"
      },
...
      "Contradiction": {
        "description": "Dos campos que no pueden ser ciertos a la vez, y la regla que lo dice.",
        "properties": {
          "fields": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "Fields",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "rule": {
            "title": "Rule",
            "type": "string"
          }
        },
        "required": [
          "fields",
          "rule",
          "message"
        ],
        "title": "Contradiction",
        "type": "object"
      },
...
      "Draft": {
        "description": "RF-169 del frontend: lo que el panel muestra. Vacio al empezar.",
        "properties": {
          "answered": {
            "default": [],
            "items": {
              "enum": [
                "title",
                "recipient.name",
                "recipient.age",
                "recipient.role",
                "premise",
                "genre",
                "tone",
                "target_words",
                "start",
                "dedication",
                "recipient.traits",
                "recipient.memories",
                "forbidden_words",
                "forbidden_themes"
              ],
              "type": "string"
            },
            "title": "Answered",
            "type": "array"
          },
          "dedication": {
            "default": "",
            "title": "Dedication",
            "type": "string"
          },
          "entities": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExtraEntity"
            },
            "title": "Entities",
            "type": "array"
          },
          "forbidden_themes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Forbidden Themes",
            "type": "array"
          },
          "forbidden_words": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Forbidden Words",
            "type": "array"
          },
          "genre": {
            "default": "",
            "title": "Genre",
            "type": "string"
          },
          "premise": {
            "default": "",
            "title": "Premise",
            "type": "string"
          },
          "recipient_age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient Age"
          },
          "recipient_memories": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Recipient Memories",
            "type": "array"
          },
          "recipient_name": {
            "default": "",
            "title": "Recipient Name",
            "type": "string"
          },
          "recipient_role": {
            "default": "",
            "title": "Recipient Role",
            "type": "string"
          },
          "recipient_traits": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Recipient Traits",
            "type": "array"
          },
          "start": {
            "default": "",
            "title": "Start",
            "type": "string"
          },
          "target_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Words"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "tone": {
            "default": "",
            "title": "Tone",
            "type": "string"
          }
        },
        "title": "Draft",
        "type": "object"
      },
...
      "ExtraEntity": {
        "description": "Un personaje o un lugar que entro al aceptar un hecho propuesto (RF-218).",
        "properties": {
          "kind": {
            "enum": [
              "person",
              "place"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "note": {
            "default": "",
            "maxLength": 300,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "name"
        ],
        "title": "ExtraEntity",
        "type": "object"
      },
...
      "FieldEdit": {
        "description": "RF-214. Un campo por nombre; los de lista, separados por comas.",
        "properties": {
          "field": {
            "enum": [
              "title",
              "recipient.name",
              "recipient.age",
              "recipient.role",
              "premise",
              "genre",
              "tone",
              "target_words",
              "start",
              "dedication",
              "recipient.traits",
              "recipient.memories",
              "forbidden_words",
              "forbidden_themes"
            ],
            "title": "Field",
            "type": "string"
          },
          "value": {
            "maxLength": 4000,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "field",
          "value"
        ],
        "title": "FieldEdit",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterviewState": {
        "description": "RI-39, RI-40 del frontend.",
        "properties": {
          "brief": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Brief"
              },
              {
                "type": "null"
              }
            ],
            "description": "En la forma de RI-01, solo si el brief esta completo"
          },
          "complete": {
            "title": "Complete",
            "type": "boolean"
          },
          "contradictions": {
            "items": {
              "$ref": "#/components/schemas/Contradiction"
            },
            "title": "Contradictions",
            "type": "array"
          },
          "discarded_quotes": {
            "description": "Citas descartadas por no aparecer literales",
            "minimum": 0.0,
            "title": "Discarded Quotes",
            "type": "integer"
          },
          "draft": {
            "$ref": "#/components/schemas/Draft"
          },
          "interview_id": {
            "title": "Interview Id",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "title": "Messages",
            "type": "array"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/Missing"
            },
            "title": "Missing",
            "type": "array"
          },
          "novel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Identificador sugerido para RI-01 (D-81)",
            "title": "Novel Id"
          },
          "proposed": {
            "items": {
              "$ref": "#/components/schemas/ProposedFact"
            },
            "title": "Proposed",
            "type": "array"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Question"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "interview_id",
          "messages",
          "question",
          "draft",
          "missing",
          "contradictions",
          "proposed",
          "discarded_quotes",
          "complete",
          "brief",
          "novel_id"
        ],
        "title": "InterviewState",
        "type": "object"
      },
...
      "Message": {
        "properties": {
          "kind": {
            "enum": [
              "question",
              "answer",
              "free_text",
              "note"
            ],
            "title": "Kind",
            "type": "string"
          },
          "role": {
            "enum": [
              "interviewer",
              "person"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "role",
          "kind",
          "text"
        ],
        "title": "Message",
        "type": "object"
      },
...
      "Missing": {
        "properties": {
          "field": {
            "enum": [
              "title",
              "recipient.name",
              "recipient.age",
              "recipient.role",
              "premise",
              "genre",
              "tone",
              "target_words",
              "start",
              "dedication",
              "recipient.traits",
              "recipient.memories",
              "forbidden_words",
              "forbidden_themes"
            ],
            "title": "Field",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "field",
          "label"
        ],
        "title": "Missing",
        "type": "object"
      },
...
      "ProposedFact": {
        "description": "RF-174 del frontend: un hecho sacado del texto libre, con su cita.",
        "properties": {
          "fact_id": {
            "title": "Fact Id",
            "type": "string"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "status": {
            "default": "proposed",
            "enum": [
              "proposed",
              "accepted",
              "discarded"
            ],
            "title": "Status",
            "type": "string"
          },
          "target": {
            "enum": [
              "recipient.traits",
              "recipient.memories",
              "entity.person",
              "entity.place"
            ],
            "title": "Target",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "fact_id",
          "target",
          "value",
          "quote"
        ],
        "title": "ProposedFact",
        "type": "object"
      },
      "Question": {
        "properties": {
          "field": {
            "enum": [
              "title",
              "recipient.name",
              "recipient.age",
              "recipient.role",
              "premise",
              "genre",
              "tone",
              "target_words",
              "start",
              "dedication",
              "recipient.traits",
              "recipient.memories",
              "forbidden_words",
              "forbidden_themes"
            ],
            "title": "Field",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "field",
          "text"
        ],
        "title": "Question",
        "type": "object"
      },
      "Recipient": {
        "description": "RF-200. Para quien es la novela: una entidad del brief, con lo que la entrevista recogio.",
        "properties": {
          "age": {
            "maximum": 120.0,
            "minimum": 0.0,
            "title": "Age",
            "type": "integer"
          },
          "entity_id": {
            "minLength": 1,
            "title": "Entity Id",
            "type": "string"
          },
          "memories": {
            "items": {
              "type": "string"
            },
            "title": "Memories",
            "type": "array"
          },
          "role": {
            "description": "Papel en la historia",
            "minLength": 1,
            "title": "Role",
            "type": "string"
          },
          "traits": {
            "items": {
              "type": "string"
            },
            "title": "Traits",
            "type": "array"
          }
        },
        "required": [
          "entity_id",
          "age",
          "role"
        ],
        "title": "Recipient",
        "type": "object"
      },
...
      "TurnIn": {
        "description": "RI-39. Todo opcional: un turno puede traer cualquier combinacion.",
        "properties": {
          "accept": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Accept",
            "type": "array"
          },
          "answer": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "discard": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Discard",
            "type": "array"
          },
          "edits": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldEdit"
            },
            "maxItems": 20,
            "title": "Edits",
            "type": "array"
          },
          "free_text": {
            "anyOf": [
              {
                "maxLength": 200000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Texto no confiable: una anecdota, una carta. Nunca instruye al sistema",
            "title": "Free Text"
          }
        },
        "title": "TurnIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorldTime": {
        "description": "MUN-05. Instante de mundo.\n\nISO 8601 mas `seq` de desempate (D-07). Se ordena por la tupla\n`(stamp, seq)`, que ordena lexicograficamente igual que cronologicamente:\npor eso el formato es fijo y se valida.",
        "properties": {
          "seq": {
            "default": 0,
            "description": "Desempate dentro del mismo instante",
            "minimum": 0.0,
            "title": "Seq",
            "type": "integer"
          },
          "stamp": {
            "description": "ISO 8601: YYYY-MM-DD, opcionalmente con hora",
            "title": "Stamp",
            "type": "string"
          }
        },
        "required": [
          "stamp"
        ],
        "title": "WorldTime",
        "type": "object"
      },
      "brief__routes__ErrorDetail": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Sistema autonomo de generacion de novelas largas. La API observa y arranca; no aprueba, no corrige y no desbloquea.",
    "title": "Story Maker",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/interviews/{interview_id}": {
      "get": {
        "description": "RI-40. El mismo estado que RI-39, para reanudar.",
        "operationId": "get_interview_interviews__interview_id__get",
        "parameters": [
          {
            "description": "Identificador de la entrevista",
            "in": "path",
            "name": "interview_id",
            "required": true,
            "schema": {
              "description": "Identificador de la entrevista",
              "pattern": "^[a-f0-9]{12}$",
              "title": "Interview Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewState"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/brief__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Interview",
        "tags": [
          "brief"
        ]
      }
    },
    "/interviews/{interview_id}/turns": {
      "post": {
        "description": "RI-39. Respuesta, texto libre, ediciones y decisiones sobre hechos propuestos.",
        "operationId": "send_turn_interviews__interview_id__turns_post",
        "parameters": [
          {
            "description": "Identificador de la entrevista",
            "in": "path",
            "name": "interview_id",
            "required": true,
            "schema": {
              "description": "Identificador de la entrevista",
              "pattern": "^[a-f0-9]{12}$",
              "title": "Interview Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewState"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/brief__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Turn",
        "tags": [
          "brief"
        ]
      }
    },
...
  }
}